+++ src/mempool/mempool_test.rs
@@ -15,7 +15,7 @@ mod tests {
         },
         storage::store::{
             engine::{MempoolMessage, ShardEngine},
-            test_helper::{self, commit_event, default_storage_event, EngineOptions, FID_FOR_TEST},
+            test_helper::{self, commit_event, default_storage_event, FID_FOR_TEST},
         },
         utils::{
             factory::{events_factory, messages_factory},
@@ -34,9 +34,8 @@ mod tests {
     const HOST_FOR_TEST: &str = "127.0.0.1";
     const PORT_FOR_TEST: u32 = 9388;
 
-    fn setup(
+    async fn setup(
         config: Option<Config>,
-        engine_options: Option<EngineOptions>,
         enable_rate_limits: bool,
     ) -> (
         ShardEngine,
@@ -57,10 +56,7 @@ mod tests {
         let (mempool_tx, mempool_rx) = mpsc::channel(100);
         let (messages_request_tx, messages_request_rx) = mpsc::channel(100);
         let (shard_decision_tx, shard_decision_rx) = broadcast::channel(100);
-        let (engine, _) = match engine_options {
-            Some(engine_options) => test_helper::new_engine_with_options(engine_options),
-            None => test_helper::new_engine(),
-        };
+        let (engine, _) = test_helper::new_engine();
         let mut shard_senders = HashMap::new();
         shard_senders.insert(1, engine.get_senders());
         let mut shard_stores = HashMap::new();
@@ -76,6 +72,7 @@ mod tests {
                     proto::FarcasterNetwork::Devnet,
                     statsd_client.clone(),
                 )
+                .await
                 .unwrap(),
             ),
             None => None,
@@ -111,7 +108,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_duplicate_user_message_is_invalid() {
-        let (mut engine, _, mut mempool, _, _, _, _) = setup(None, None, false);
+        let (mut engine, _, mut mempool, _, _, _, _) = setup(None, false).await;
         test_helper::register_user(
             1234,
             default_signer(),
@@ -130,7 +127,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_duplicate_onchain_event_is_invalid() {
-        let (mut engine, _, mut mempool, _, _, _, _) = setup(None, None, false);
+        let (mut engine, _, mut mempool, _, _, _, _) = setup(None, false).await;
         let onchain_event = events_factory::create_rent_event(1234, Some(10), None, false);
         let (valid, _) = mempool.message_is_valid(&MempoolMessage::ValidatorMessage(ValidatorMessage {
             on_chain_event: Some(onchain_event.clone()),
@@ -147,7 +144,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_duplicate_fname_transfer_is_invalid() {
-        let (mut engine, _, mut mempool, _, _, _, _) = setup(None, None, false);
+        let (mut engine, _, mut mempool, _, _, _, _) = setup(None, false).await;
         test_helper::register_user(
             1,
             default_signer(),
@@ -182,7 +179,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_rate_limits_applied() {
-        let (mut engine, _, mut mempool, _, _, _, _) = setup(None, None, true);
+        let (mut engine, _, mut mempool, _, _, _, _) = setup(None, true).await;
 
         let id_register_event = events_factory::create_id_register_event(
             FID_FOR_TEST,
@@ -212,7 +209,8 @@ mod tests {
 
     #[tokio::test]
     async fn test_mempool_size() {
-        let (_, _, mut mempool, mempool_tx, _request_tx, _decision_tx, _) = setup(None, None, false);
+        let (_, _, mut mempool, mempool_tx, _request_tx, _decision_tx, _) =
+            setup(None, false).await;
         tokio::spawn(async move {
             mempool.run().await;
         });
@@ -247,7 +245,7 @@ mod tests {
     #[tokio::test]
     async fn test_mempool_prioritization() {
         let (_, _, mut mempool, mempool_tx, messages_request_tx, _shard_decision_tx, _) =
-            setup(None, None, false);
+            setup(None, false).await;
 
         // Spawn mempool task
         tokio::spawn(async move {
@@ -323,7 +321,7 @@ mod tests {
     #[tokio::test]
     async fn test_mempool_eviction() {
         let (mut engine, _, mut mempool, mempool_tx, messages_request_tx, shard_decision_tx, _) =
-            setup(None, None, false);
+            setup(None, false).await;
         test_helper::register_user(
             1234,
             default_signer(),
@@ -417,7 +415,7 @@ mod tests {
         let config2 = Config::new(node2_addr.clone(), node1_addr.clone());
 
         let (_, gossip1, mut mempool1, mempool_tx1, _mempool_requests_tx1, _shard_decision_tx1, _) =
-            setup(Some(config1), None, false);
+            setup(Some(config1), false).await;
         let (
             _,
             gossip2,
@@ -426,7 +424,7 @@ mod tests {
             mempool_requests_tx2,
             _shard_decision_tx1,
             mut system_rx2,
-        ) = setup(Some(config2), None, false);
+        ) = setup(Some(config2), false).await;
 
         // Spawn gossip tasks
         tokio::spawn(async move {
@@ -515,13 +513,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_rate_limits_exceeded() {
-        let limits = test_helper::limits::test_store_limits();
-        let engine_options = test_helper::EngineOptions {
-            limits: Some(limits.clone()),
-            db: None,
-            messages_request_tx: None,
-        };
-        let (mut engine, _, mut mempool, _, _, _, _) = setup(None, Some(engine_options), true);
+        let (mut engine, _, mut mempool, _, _, _, _) = setup(None, true).await;
         test_helper::register_user(
             FID_FOR_TEST,
             default_signer(),